Expand classes for doc generation
