+++ docs/.vuepress/config.js
@@ -12,6 +12,7 @@ module.exports = {
       { text: 'GitHub', link: 'https://github.com/polkadot-js/ui' }
     ],
     sidebar: [
+      ['/ui-identicon/', '@polkadot/ui-identicon'],
       ['/ui-keyring/', '@polkadot/ui-keyring'],
       ['/ui-settings/', '@polkadot/ui-settings'],
       ['/ui-util/', '@polkadot/ui-util'],
