+++ docs/README.md
@@ -1,0 +1,21 @@
+# Polkadot Javascript UI libraries
+
+This is a collection of UI (Rwact) libraries, utlities and classes that allows for easier use of the Polkadot JS infrstaructure inside your apps.
+
+## The Polkadot Project
+
+You can read more about the Polkadot Network at [https://polkadot.network/](https://polkadot.network/) and more about the polkadot-js projects at [https://polkadot.js.org](https://polkadot.js.org)
+
+## Github repositories
+
+You can find the Polkadot repositories at:
+
+- [https://github.com/polkadot-js](https://github.com/polkadot-js)
+- [https://github.com/paritytech/polkadot](https://github.com/paritytech/polkadot)
+- [https://github.com/paritytech/substrate](https://github.com/paritytech/substrate)
+
+This documentation is generated from [https://github.com/polkadot-js/ui](https://github.com/polkadot-js/ui)
+
+## Contributing
+
+Contribution to the Polkadot JS projects are more than welcome. You can [report issues](https://github.com/polkadot-js/ui/issues/new) and [log feature requests](https://github.com/polkadot-js/ui/issues/new).
