+++ packages/ui-identicon/src/index.ts
@@ -1,0 +1,7 @@
+// Copyright 2017-2019 @polkadot/ui-identicon authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import { default as Identicon } from './Identicon';
+
+export default Identicon;
