+++ typedoc.js
@@ -1,0 +1,16 @@
+module.exports = {
+  name: 'Polkadot JS UI libraries',
+  exclude: '**/*+(index|e2e|spec).ts',
+  excludeExternals: true,
+  excludeNotExported: true,
+  excludeProtected: true,
+  excludePrivate: true,
+  hideGenerator: true,
+  includeDeclarations: false,
+  out: 'docs',
+  module: 'commonjs',
+  moduleResolution: 'node',
+  mdEngine: 'gitbook',
+  stripInternal: 'true',
+  theme: 'markdown'
+};
